Add judgment logic on whether it is installed in the installer script #21

+++ install/MacOSX/10_create_centos6-64bit_box.sh
@@ -4,14 +4,21 @@ os_name="centos6-64bit"
 box_url="https://github.com/tommy-muehle/puppet-vagrant-boxes/releases/download/1.0.0/centos-6.6-x86_64.box"
 
 vagrant plugin install vagrant-hostsupdater
-vagrant init ${os_name}
-vagrant box add ${os_name} ${box_url}
 
-rm Vagrantfile
-mv .vagrant/ ../../centos6-64bit/
-cd ../../centos6-64bit/
+if [  -e ../../centos6-64bit/.vagrant ]; then
+    vagrant init ${os_name}
+    vagrant box add ${os_name} ${box_url}
 
-echo " Completion of virtual box creation."
+    rm Vagrantfile
+    mv .vagrant/ ../../centos6-64bit/
+    echo ""
+    echo " Completion of virtual box creation."
+else
+    echo ""
+    echo " Already exists.No change."
+    echo " Or delete the \". Vagrant\" folder."
+fi
+
 echo ""
 echo "==== Execute these commands. ==="
 echo ""
